mode edit

+++ src/main/java/servlet/finalexam.java
@@ -108,10 +108,10 @@ public void doPost (HttpServletRequest request, HttpServletResponse response)
    else if (operation.equals(OperationMode))
    {
       double mode = array.get(0);
-        int maxCount = 0;
+        double maxCount = 0.0;
         for (int i = 0; i < array.size(); i++) {
             double value = array.get(i);
-            int count = 0;
+            double count = 0.0;
             for (int j = 0; j < array.size(); j++) {
                 if (array.get(j) == value) count++;
                 if (count > maxCount) {
